requirements.txt
#requirements.txt
psycopg2-binary==2.9.7
streamlit-extras==0.3.2
streamlit-authenticator==0.2.2
bcrypt==4.0.1